PKEntity Deleter Update
There was an issue where the server and client mismatched what was deleted. Suspected cause was when the event was executed EonInit , This has been changed to EonActivate this has solved the issue with Peer testing the client is not being kicked. This being said I have really no clue what I have done.